+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [v1.0.1]
+
+### Updated
+
++ composer settings
+
 ## [v1.0.0]
 
 ### Added
@@ -25,6 +31,7 @@
 + Facility model and repo 
 + add Last changes endpoint
 
+[v1.0.1]: https://github.com/towa-digital/booking-com-sdk/releases/tag/v1.0.1
 [v1.0.0]: https://github.com/towa-digital/booking-com-sdk/releases/tag/v1.0.0
 [v0.0.2]: https://github.com/towa-digital/booking-com-sdk/releases/tag/v0.0.2
 [unreleased]: https://github.com/towa-digital/booking-com-sdk/tree/develop